Use globals to reduce argument count for functions
Store parsing option values in global variables rather than passing them
as arguments to every function that needs them. getparams was originally
intended to be sourced rather than executed. This meant only creating
global variables intended for use in the sourcing script.

Now that getparams is a command rather than a library, complexity of the
script can be reduced through the use of global variables.