Output option arguments as separate parameters by default (#4)
BREAKING CHANGE: Use the behavior of the --no-combine-args option
as the default behavior.

Currently, options with arguments are output as --option='arg' and
-o='arg'. This allows differentiation between empty and unspecified
optional arguments, but puts a slightly bigger burden on scripts with
options that take arguments and is unconventional (and possibly
unintuitive) behavior.

Output arguments as the next parameter following their corresponding
options by default. Empty and unspecified optional arguments are both
output as a '' parameter.

Users used to the getopt(1) command will more or less expect this as
the default behavior and most use-cases probably don't call for
differentiation between unspecified and empty optional arguments.