Add test dependency bats-core as submodule
Adds bats-core as a submodule to remove the need for BATS to be manually
installed on a machine in order to run the test suite.

Also removes 'Setup BATS' step from 'test' job in the CI workflow as the
submodule is automatically initialized (if not already) when running the
'test' make target.